pyproject.toml
[project]
name = "hg-final-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.32.0",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.2.0",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "supabase>=2.15.2",
]
